backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
weaviate-client==4.16.4
chromadb==0.4.22
sentence-transformers==2.2.2
openai==1.30.1
httpx==0.27.0
requests==2.31.0
numpy== 1.24.3